Move `controllerdeployment` controller
Preparation to introduce a `reference` controller
